data store working and login bypassed if has data inside
