+++ app/src/main/java/com/ort/edu/parcialtp3/MainActivity.kt
@@ -1,9 +1,7 @@
 package com.ort.edu.parcialtp3
 
 import android.content.Context
-import android.content.Intent
 import android.os.Bundle
-import android.view.MenuItem
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.GravityCompat
@@ -15,7 +13,6 @@ import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.NavigationUI
 import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.navigation.NavigationView
-import com.ort.edu.parcialtp3.UserSession.userName
 
 val Context.dataStore by preferencesDataStore(name = "USER_PREFERENCES_NAME")
 
@@ -24,17 +21,14 @@ class MainActivity : AppCompatActivity() {
     private lateinit var navHostFragment : NavHostFragment
     private lateinit var drawerLayout : DrawerLayout
     private lateinit var navigationView : NavigationView
+    private lateinit var name: String
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
         drawerLayout = findViewById(R.id.drawer_layout)
         navigationView = findViewById(R.id.nav_view)
         navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
-//        val navController = navHostFragment.navController
-//        findViewById<NavigationView>(R.id.nav_view)
-//            .setupWithNavController(navController)
 
         setUpDrawerLayout()
