+++ app/src/main/java/com/ort/edu/parcialtp3/fragments/home.kt
@@ -5,6 +5,7 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.TextView
 import androidx.appcompat.widget.SearchView
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.fragment.app.Fragment
@@ -22,6 +23,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -56,6 +58,7 @@ class HomeFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
         characterRecyclerView = view.findViewById(R.id.characterRecyclerView)
         characterList = arrayListOf<Character>()
+        val textHome = view.findViewById<TextView>(R.id.homeTextView)
 
         getCharacters()
         // Configuro el recyclerview y le paso un Adapter
@@ -78,20 +81,9 @@ class HomeFragment : Fragment() {
                 return false
             }
         })
-
-        lifecycleScope.launch(Dispatchers.IO){
-            getUserData().collect{
-                searchView.
-            }
-        }
     }
 
-    private fun getUserData() = requireContext().dataStore.data.map { preferences ->
-        UserData(
-            name = preferences[stringPreferencesKey("name")].orEmpty(),
-            password = preferences[stringPreferencesKey("password")].orEmpty()
-        )
-    }
+
 
     fun getCharacters() {
         val service = ApiServiceBuilder.create()
